workflows: Only run the CI action on pull_request events
Both the CI and CD workflows are defined to be executed on the push event.
According to the Github documentation, in that case they are executed in
the order that were commited.

Even though the CI action was commited before the CD action and so they
will be executed in the correct order, it's better to make it explicit.

One option is to use dispatch events and waiting for events, but that is
overkill for the current pipeline, instead just execute the CI only for
the pull_request event and the CD only for the push event. After all, the
changes should also be made via pull requests and never pushed directly.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>